Configure default platform and configuration (#20)
Rather than requiring customers to set a default configuration and platform to get output to be consistent, I've defaulted them to `Debug` and `AnyCPU` respectively. As an option, you can set `CentralBuildOutputNoDefaultPlatform` to `true` to allow omitting the platform in the output in the case of the default `AnyCPU` platform.