2.0 (#73)
- Lots of engineering improvements and efforts to make things consistent
with some of my other projects. This includes code style changes and
updating the Dependabot schedule.
- Changed and documented the release mechanics.
- BREAKING: Removed test validation using .NET Core 3.1 as it is now out
of support. This does not mean that compatibility with .NET Core 3.1 has
been broken since I haven't changed anything specific to .NET Core 3.1,
but an effort will not be made to ensure compatibility going forward.
This necessitated a bump to 2.0 (do release).