Resolve symlink on macOS differently (#131)
Previously, I had used a library (`Mono.Posix.NETStandard`) to resolve
the symlink to the temp path on macOS to get a path that I could compare
against after retrieving values from MSBuild. Recently, the default
build agents in GitHub were updated to use newer ARM64 machines, which
that library does not support. I attempted to use
[Directory.ResolveLinkTarget](https://learn.microsoft.com/dotnet/api/system.io.directory.resolvelinktarget)
(available as of .NET 6) to resolve the path, but it says it's not a
symlink. I'm left with this `/private` prefix. Instead, I am going to
use the
[RUNNER_TEMP](https://docs.github.com/actions/learn-github-actions/variables#default-environment-variables)
environment variable on runners to get a more stable temporary location.
Added some logging for future reference.